+++ Runtime/UI.cs
@@ -13,6 +13,11 @@ namespace Pixygon.Micro {
         [SerializeField] private TextMeshPro _pointsText;
         [SerializeField] private TextMeshPro _levelStartText;
 
+        [SerializeField] private bool _animatePoints;
+        [SerializeField] private bool _animateCoins;
+        [SerializeField] private Animator _pointsAnimator;
+        [SerializeField] private Animator _coinsAnimator;
+        
         public GameObject PregameScreen => _pregameScreen;
         public GameObject LevelEndScreen => _levelEndScreen;
         public void SetLife(int i) {
@@ -22,10 +27,12 @@ namespace Pixygon.Micro {
         }
 
         public void SetCoins(int coins) {
+            if(_animateCoins) _coinsAnimator.SetTrigger("Tick");
             _coinText.text = coins.ToString();
         }
 
         public void SetPoints(int points) {
+            if(_animatePoints) _pointsAnimator.SetTrigger("Tick");
             _pointsText.text = points.ToString();
         }
         public void TriggerMenuScreen(bool activate) {